udbm-backend/requirements.txt
# Web框架
fastapi
uvicorn[standard]

# 数据库ORM
sqlalchemy[asyncio]
asyncpg  # PostgreSQL异步驱动
psycopg2-binary

# 数据验证
pydantic
pydantic-settings

# 安全认证
python-jose[cryptography]
passlib[bcrypt]

# HTTP客户端
httpx

# 工具库
python-multipart
python-dotenv

# 监控和日志
prometheus-client
structlog

# 开发工具
pytest
pytest-asyncio

# 额外依赖
greenlet
typing-extensions
psutil
